Split bus.c and device.c. Rename a function.
