+++ sys/Makefile
@@ -4,6 +4,7 @@ TOPDIR = $(realpath ..)
 
 SOURCES_C  = \
 	assert.c \
+	bus.c \
 	callout.c \
 	clock.c \
 	condvar.c \
