+++ sys/device.c
@@ -1,8 +1,6 @@
 #define KL_LOG KL_DEV
-#include <stdc.h>
 #include <klog.h>
 #include <device.h>
-#include <pci.h>
 #include <rman.h>
 
 MALLOC_DEFINE(M_DEV, "devices & drivers", 128, 1024);
@@ -47,30 +45,18 @@ int device_detach(device_t *dev) {
   return res;
 }
 
-int bus_generic_probe(device_t *bus) {
-  device_t *dev;
-  SET_DECLARE(driver_table, driver_t);
-  klog("Scanning %s for known devices.", bus->driver->desc);
-  TAILQ_FOREACH (dev, &bus->children, link) {
-    driver_t **drv_p;
-    SET_FOREACH(drv_p, driver_table) {
-      driver_t *drv = *drv_p;
-      dev->driver = drv;
-      if (device_probe(dev)) {
-        klog("%s detected!", drv->desc);
-        device_attach(dev);
-        break;
-      }
-      dev->driver = NULL;
-    }
-  }
-  return 0;
-}
-
 device_t *make_device(device_t *parent, driver_t *driver) {
   device_t *dev = device_add_child(parent);
   dev->driver = driver;
   if (device_probe(dev))
     device_attach(dev);
   return dev;
+}
+
+void device_add_resource(device_t *dev, resource_t *r, int rid,
+                         bus_space_t *bus_space) {
+  r->r_owner = dev;
+  r->r_id = rid;
+  r->r_bus_space = bus_space;
+  LIST_INSERT_HEAD(&dev->resources, r, r_device);
 }
